Add DirectorScraper
Add DirectorScraper class. Similar to movie/celebrity, it can take the options "director_name" or "director_url". If the "print" option is set to "true", it will also print to console while it processes. The metadata returned consists of the movie title, year, RT score, and box office (in dollars) for each movie that the person directed.